*Added DVD Drive support (thanks joe (ftpii) for his libfst and libiso) You can copy files/directories from a normal DVD or a WiiDisk now. Just insert a disk and click the DeviceMenu Button.
*Added the option to load the config file from the USB Device
The app will try to find a config file on SD Card first if none is found it will try finding one on the FAT/FAT32 USB Device. If both searches failed the app will try to look from which device it was started and if that failes the default to SD is set. I couldn't test it completly because of lack of time but i hope some other people can.