Well after looking for a while into the mupdf library (thanks to the producers!) and modifying it in some parts here is something new:
New Feature: PDF Viewer :D

This makes WiiXplorer the first PDF Viewer for the Wii (i think :P, didn't see any others)

NOTE 1:
The navigations of the new PDF Viewer is the same as the Image Viewer. In the Settings you can configure a "PDF Processing Zoom" which is used for internal PDF Processing. With this you can improve quality because the PDF is processed in a higher resolution (very usefull sometimes). It is different than the Zoom in the PDF Viewer. Be aware that the Wii does not have endless memory and can't handle too high resolutions or some PDFs at all. The GPU limit is at about 1024x1024.

NOTE 2: The PDF Viewer uses some fonts for some PDFs. You can find a bunch in the download section. They go into the "UpdatePath/fonts/" which is default to "sd:/apps/WiiExplorer/fonts/"

NOTE 3: For the Japanese/Chinese and some more Users probably that use special signs, a CJK support is sometimes needed which processes CMaps. I compiled the CMaps and included them to the source BUT I did not include them into the app because it adds about 3 MB to the app. If you want to activate the feature here is a little tutorial how to:

1. Open the Makefile and search for "-DNOCJK"
2. Remove this from the Makefile
3. Recompile (clean/rebuild) the complete source

The app should be around 6-7MB big. Be aware that memory is a big issue in PDF processing and you might not even be able to open some files.