*Language file update *Fix for issue 451
New file systems support: EXT2/EXT3/EXT4
I ported libext2fs for the Wii/GC for all the Linux lovers out there :P. It is based on the source of libntfs and is build up in a similar way.
The source code for the port can be found under this SVN repository:
http://code.google.com/p/libext2fs-wii/

This is an initial release for the libext2fs-wii port (v1.0.0). Even though I did several tests by now, it might still have bugs/issues in it. Until it is further tested use with cautious and on your own risk, especially the write/delete feature. 
I will create a WiiBrew page for it later. If someone wants to go ahead and make one, feel free to do so :P (I am a bit low on free time currently).